+++ docs/app/_components/PortalContainer.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-export const PortalContaier = () => {
-  return (
-    <div data-theme="rui">
-      <div id="portalContainer" className="not-prose" />
-    </div>
-  );
-};