+++ packages/components/src/Table/Table.tsx
@@ -116,8 +116,12 @@ export const Table: Table = ({
       <table
         ref={tableRef}
         className={cn(
-          'group/table',
-          'border-collapse',
+          'group/table border-collapse',
+          /**
+           * Prevents wide tables from causing overlays to become scrollable on
+           * small screens, ensuring overlays remain fixed as intended.
+           */
+          'max-[600px]:[[aria-hidden="true"]_&]:overflow-hidden',
           stretch ? 'table w-full' : 'block',
           classNames.table
         )}